Adding validations.
This commit introduces a client validator framework and a validated
input directive. The validator takes a RAML named parameter description
and generates a Validator from it. The validated-input takes a RAML
named parameter description, creates a validator and then applies the
validation when the input loses focus (ie on blur).